+++ CHANGES.md
@@ -30,6 +30,7 @@
     * Fixed configuration panel detection of modification.
     * Fixed an issue where an invalid password would be considered needing to log in
       again with the existing, known password.
+    * Stream-lined the server connection process.
     * Temporarily disabled the client name refresh button until that's all worked out.